+++ src/c_cmds.c
@@ -505,6 +505,8 @@ consolecmd_t consolecmds[] =
         "Dumps the console to a file."),
     CVAR_INT(crosshair, "", crosshair_cvar_func1, crosshair_cvar_func2, CF_NONE, CROSSHAIRVALUEALIAS,
         "Toggles displaying a crosshair when using mouselook\n(<b>none</b>, <b>cross</b> or <b>dot</b>)."),
+    CVAR_INT(crosshaircolor, crosshaircolour, color_cvars_func1, color_cvars_func2, CF_NONE, NOVALUEALIAS,
+        "The color of the crosshair (<b>0</b> to <b>255</b>, or <b>#</b><i>rrggbb</i>)."),
     CMD(cvarlist, "", null_func1, cvarlist_cmd_func2, true, "[<i>searchstring</i>]",
         "Lists all console variables."),
     CMD(endgame, "", game_func1, endgame_cmd_func2, false, "",